Removed the first-time-eligibility-file and unit test. It's contents have now been split out into separate files with separate unit tests. Moved views and routes under first time directory. Removed completed TODOs.
