+++ app/views/first-time/benefits.html
@@ -26,7 +26,7 @@
             {% endif %}
 
             <form action="/first-time/{{ dob }}/{{ relationship }}/{{ journeyAssistance }}" method="post">
-                <div class="form-group {% if errors['benefit'][0] %} error {% endif %}"">
+                <div class="form-group {% if errors['benefit'][0] %} error {% endif %}">
                     <fieldset>
 
                         <label for="income-support" class="block-label">
